+++ src/main/java/uk/ashleybye/tictactoe/ui/console/ConsolePlayerConfiguration.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe.console;
+package uk.ashleybye.tictactoe.ui.console;
 
 import uk.ashleybye.tictactoe.core.Mark;
 import uk.ashleybye.tictactoe.core.PlayerConfiguration;
