+++ src/main/java/uk/ashleybye/tictactoe/ui/console/ConsolePlayerFactory.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe.console;
+package uk.ashleybye.tictactoe.ui.console;
 
 import java.util.Arrays;
 import java.util.List;
