+++ src/test/java/uk/ashleybye/tictactoe/ui/console/MockGameReport.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe.console;
+package uk.ashleybye.tictactoe.ui.console;
 
 import java.util.HashMap;
 import java.util.Map;
